+++ app/assets/javascripts/hyrax.js
@@ -37,12 +37,9 @@
 //= require webcomponentsjs/0.5.4/CustomElements.min
 //= require time-elements
 
-// ActionCable infrastructure
-// NOTE: This should be generated by Rails into downstream apps -- should this live here in Hyrax?
-//= require cable
-
 //= require hyrax/monkey_patch_turbolinks
 //= require hyrax/app
+//= require hyrax/cable
 //= require hyrax/config
 //= require hyrax/initialize
 //= require hyrax/fileupload
