+++ app/assets/javascripts/hyrax/cable.js
@@ -8,5 +8,5 @@
 (function() {
   this.App || (this.App = {});
 
-  App.cable = ActionCable.createConsumer();
+  App.cable = ActionCable.createConsumer("/notifications/endpoint");
 }).call(this);
