+++ app/assets/javascripts/hyrax/channels/notifications.js
@@ -1,22 +1,26 @@
 App.cable.subscriptions.create("Hyrax::NotificationsChannel", {
     received: function(data) {
-        let count = data["notify_number"]
-        let notification = $('.notify_number');
+        console.log("received");
+        let count = data["notify_number"];
+        let notification = $('.notify-number');
         let countSelector = notification.find('.count');
         countSelector.html(count);
         if (count == 0) {
+            console.log("count is zero");
             countSelector.addClass('invisible');
             notification.prop('aria-label', 'You have no unread notifications');
         }
         else if (count == 1) {
+            console.log("count is one");
             countSelector.removeClass('invisible');
             notification.prop('aria-label', 'You have one unread notification');
-            notification.addClass('label-danger').removeClass('label-default');
+            countSelector.addClass('label-danger').removeClass('label-default');
         }
         else {
+            console.log("count is else");
             countSelector.removeClass('invisible');
-            notification.prop('aria-label', 'You have %{count} unread notifications');
-            notification.addClass('label-danger').removeClass('label-default');
+            notification.prop('aria-label', 'You have ' + count.toString() + ' unread notifications');
+            countSelector.addClass('label-danger').removeClass('label-default');
         }
     }
 });
