--- app/channels/application_cable/connection.rb
@@ -1,28 +1,0 @@
-module ApplicationCable
-  class Connection < ActionCable::Connection::Base
-    identified_by :current_user
-
-    def connect
-      self.current_user = find_verified_user
-    end
-
-    private
-
-      def find_verified_user
-        user = ::User.find_by(id: user_id)
-        if user
-          user
-        else
-          reject_unauthorized_connection
-        end
-      end
-
-      def user_id
-        session['warden.user.user.key'][0][0]
-      end
-
-      def session
-        cookies.encrypted[Rails.application.config.session_options[:key]]
-      end
-  end
-end
