+++ app/channels/hyrax/application_cable/channel.rb
@@ -1,0 +1,6 @@
+module Hyrax
+  module ApplicationCable
+    class Channel < ActionCable::Channel::Base
+    end
+  end
+end
