+++ app/views/hyrax/users/_notify_number.html.erb
@@ -2,5 +2,5 @@
             'aria-label' => t("hyrax.toolbar.notifications"),
             class: "notify-number" do %>
     <span class="fa fa-bell"></span>
-    <span class="count invisible label label-danger">0</span>
+    <span class="count invisible label label-danger"><%= current_user.mailbox.inbox(unread: true).count %></span>
 <% end %>
